project_ejemplo/flask/requirements.txt
Flask==2.1.1
Werkzeug==2.3.7
flask-wtf
flask_login
pandas
numpy
flask-bootstrap
email_validator
flask-session
mysql-connector-python
